Adds hour picker view component
Adds a custom hour picker view component for selecting hours and periods (AM/PM).
This component uses a `UICollectionView` with custom layout for a smooth scrolling experience.
It also includes a delegate protocol to notify the parent view when the selected hour changes.

The component provides visual feedback by scaling and changing the alpha of the hour labels based on their proximity to the center of the view. The component also updates an adjacent label with the selected period.

Removes the old `UIPickerView` delegate/datasource from `StartsAtViewController` and implements the `HourPickerViewDelegate` to get updates of selected hour.

Sets height constraint to show exactly 3 cells in the view.